Major progress to images on s3
